+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ### Changed
 
+* Better heuristics for guessing text or binary file format.
+
 ### Removed
 
 ## [0.1.0] - 2022-07-28